Refactored Renderer and OpenGLContext
Renderer has been moved to the Engine project, and its rendering code has been replaced with a shader. Renderer is now a simple class that sets up a Framebuffer which can be enabled, causing all draw commands to be done to that framebuffer. The renderer's shader can then draw the contents of the framebuffer.